refactor(stt): add shared nemo/ package; decouple nemotron_asr from parakeet
Move alignment.py out of parakeet/ into a new shared mlx_audio/stt/models/nemo/
family package, and point both nemotron_asr and parakeet at it. This removes the
nemotron_asr -> parakeet sibling dependency (neither model owns the other).

Mirrors Blaizzy/mlx-audio-swift#197. Pure move, no behavior change.
